// src/interfaces/gistResponse.ts
// Generated by https://quicktype.io

export interface Award {
  title: string
  date: string
  awarder: string
  summary: string
}

export interface Location {
  address: string
  postalCode: string
  city: string
  countryCode: string
  region: string
}

export interface Profile {
  network: string
  username: string
  url: string
}

export interface Basics {
  name: string
  label: string
  image: string
  email: string
  phone: string
  phoneUrl: string
  url: string
  summary: string
  location: Location
  profiles: Profile[] | []
}

export interface Certificate {
  name: string
  issuer: string
  date: string
  url: string
}

export interface Education {
  institution: string
  area: string
  url: string
  studyType: string
  startDate: string
  endDate: string
  score: string
  degree: string
}

export interface Language {
  language: string
  fluency: string
}

export interface Skill {
  name: string
  level: string
  keywords: string[]
}

export interface Work {
  name: string
  position: string
  startDate: string
  highlights: string[]
  summary: string
  url: string
  location: string
  endDate?: string
}

export interface GistResponse {
  basics: Basics | null
  work: Work[] | []
  education: Education[] | []
  awards: Award[] | []
  certificates: Certificate[] | []
  skills: Skill[] | []
  languages: Language[] | []
}